attempting cleaner impl
